requirements.txt
alembic==1.13.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bidict==0.22.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distlib==0.3.8
docopt==0.6.2
executing==2.0.1
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipenv==2023.12.0
pipreqs==0.4.13
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-engineio==4.8.2
python-socketio==5.11.0
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0
yarg==0.1.9
zipp==3.17.0
